requirements.txt
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
pyyaml==6.0.1
flask==3.0.0
requests==2.31.0
pandas==2.0.3
scikit-learn==1.3.0